Fix an issue when counting searches
There was a mistake due to a misunderstanding of the Chrome API, but
with harmful consequences: the count of every query was the total number
of times it has been performed all-time, and not only in the requested
time range. This was fixed by adding further API calls to
`history.getVisits` (resulting in an N+1 pattern, but it seems we do
not have much of a choice).